requirements.txt
pyserial>=3.5
pillow>=10.0.0
numpy>=1.24.0
